backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
motor==3.6.0
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
transformers==4.43.3
Pillow==10.4.0
scikit-learn==1.3.0
